tests/oracle/requirements.txt
pytest>=8.0
pyyaml>=6.0
Pillow>=10.0
markdownify>=0.13
croniter>=2.0
